Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-values/line-break-ch-unit-expected.txt


FAIL 3ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 4ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 5ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 6ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 7ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 8ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 9ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 10ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 11ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 12ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 13ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 14ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 15ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 16ch should fit
FAIL 17ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 18ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 19ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 20ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 21ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 22ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 23ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 24ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 25ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 26ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 27ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 28ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 29ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 30ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 31ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 32ch should fit
FAIL 33ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 34ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 35ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 36ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 37ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 38ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 39ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 40ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 41ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 42ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 43ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 44ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 45ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 46ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 47ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 48ch should fit
FAIL 49ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 50ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 51ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 52ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 53ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 54ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 55ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 56ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 57ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 58ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 59ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 60ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 61ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 62ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 63ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 64ch should fit
FAIL 65ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 66ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 67ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 68ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 69ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 70ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 71ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 72ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 73ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 74ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 75ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 76ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 77ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 78ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 79ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 80ch should fit
FAIL 81ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 82ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 83ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 84ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 85ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 86ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 87ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 88ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 89ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 90ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 91ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 92ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 93ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 94ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 95ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 96ch should fit
FAIL 97ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 98ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
FAIL 99ch should fit assert_approx_equals: expected 10 +/- 1 but got 20
PASS 3ch should wrap
PASS 4ch should wrap
PASS 5ch should wrap
PASS 6ch should wrap
PASS 7ch should wrap
PASS 8ch should wrap
PASS 9ch should wrap
PASS 10ch should wrap
PASS 11ch should wrap
PASS 12ch should wrap
PASS 13ch should wrap
PASS 14ch should wrap
PASS 15ch should wrap
PASS 16ch should wrap
PASS 17ch should wrap
PASS 18ch should wrap
PASS 19ch should wrap
PASS 20ch should wrap
PASS 21ch should wrap
PASS 22ch should wrap
PASS 23ch should wrap
PASS 24ch should wrap
PASS 25ch should wrap
PASS 26ch should wrap
PASS 27ch should wrap
PASS 28ch should wrap
PASS 29ch should wrap
PASS 30ch should wrap
PASS 31ch should wrap
PASS 32ch should wrap
PASS 33ch should wrap
PASS 34ch should wrap
PASS 35ch should wrap
PASS 36ch should wrap
PASS 37ch should wrap
PASS 38ch should wrap
PASS 39ch should wrap
PASS 40ch should wrap
PASS 41ch should wrap
PASS 42ch should wrap
PASS 43ch should wrap
PASS 44ch should wrap
PASS 45ch should wrap
PASS 46ch should wrap
PASS 47ch should wrap
PASS 48ch should wrap
PASS 49ch should wrap
PASS 50ch should wrap
PASS 51ch should wrap
PASS 52ch should wrap
PASS 53ch should wrap
PASS 54ch should wrap
PASS 55ch should wrap
PASS 56ch should wrap
PASS 57ch should wrap
PASS 58ch should wrap
PASS 59ch should wrap
PASS 60ch should wrap
PASS 61ch should wrap
PASS 62ch should wrap
PASS 63ch should wrap
PASS 64ch should wrap
PASS 65ch should wrap
PASS 66ch should wrap
PASS 67ch should wrap
PASS 68ch should wrap
PASS 69ch should wrap
PASS 70ch should wrap
PASS 71ch should wrap
PASS 72ch should wrap
PASS 73ch should wrap
PASS 74ch should wrap
PASS 75ch should wrap
PASS 76ch should wrap
PASS 77ch should wrap
PASS 78ch should wrap
PASS 79ch should wrap
PASS 80ch should wrap
PASS 81ch should wrap
PASS 82ch should wrap
PASS 83ch should wrap
PASS 84ch should wrap
PASS 85ch should wrap
PASS 86ch should wrap
PASS 87ch should wrap
PASS 88ch should wrap
PASS 89ch should wrap
PASS 90ch should wrap
PASS 91ch should wrap
PASS 92ch should wrap
PASS 93ch should wrap
PASS 94ch should wrap
PASS 95ch should wrap
PASS 96ch should wrap
PASS 97ch should wrap
PASS 98ch should wrap
PASS 99ch should wrap
0 0
0 00
0 000
0 0000
0 00000
0 000000
0 0000000
0 00000000
0 000000000
0 0000000000
0 00000000000
0 000000000000
0 0000000000000
0 00000000000000
0 000000000000000
0 0000000000000000
0 00000000000000000
0 000000000000000000
0 0000000000000000000
0 00000000000000000000
0 000000000000000000000
0 0000000000000000000000
0 00000000000000000000000
0 000000000000000000000000
0 0000000000000000000000000
0 00000000000000000000000000
0 000000000000000000000000000
0 0000000000000000000000000000
0 00000000000000000000000000000
0 000000000000000000000000000000
0 0000000000000000000000000000000
0 00000000000000000000000000000000
0 000000000000000000000000000000000
0 0000000000000000000000000000000000
0 00000000000000000000000000000000000
0 000000000000000000000000000000000000
0 0000000000000000000000000000000000000
0 00000000000000000000000000000000000000
0 000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00
0 000
0 0000
0 00000
0 000000
0 0000000
0 00000000
0 000000000
0 0000000000
0 00000000000
0 000000000000
0 0000000000000
0 00000000000000
0 000000000000000
0 0000000000000000
0 00000000000000000
0 000000000000000000
0 0000000000000000000
0 00000000000000000000
0 000000000000000000000
0 0000000000000000000000
0 00000000000000000000000
0 000000000000000000000000
0 0000000000000000000000000
0 00000000000000000000000000
0 000000000000000000000000000
0 0000000000000000000000000000
0 00000000000000000000000000000
0 000000000000000000000000000000
0 0000000000000000000000000000000
0 00000000000000000000000000000000
0 000000000000000000000000000000000
0 0000000000000000000000000000000000
0 00000000000000000000000000000000000
0 000000000000000000000000000000000000
0 0000000000000000000000000000000000000
0 00000000000000000000000000000000000000
0 000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
0 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
